+++ crates/js-sys/tests/wasm/Date.rs
@@ -484,7 +484,11 @@ fn to_locale_string() {
 fn to_locale_time_string() {
     let date = Date::new(&"August 19, 1975 23:15:30".into());
     assert_eq!(
-        JsValue::from(date.to_locale_time_string("en-US", &JsValue::undefined())),
+        JsValue::from(date.to_locale_time_string("en-US")),
+        "11:15:30 PM",
+    );
+    assert_eq!(
+        JsValue::from(date.to_locale_time_string_with_options("en-US", &JsValue::undefined())),
         "11:15:30 PM",
     );
 }