+++ crates/js-sys/CHANGELOG.md
@@ -3,9 +3,9 @@
 --------------------------------------------------------------------------------
 ## Unreleased
 
-### Changed
+### Added
 
-* Added `options` parameter to `Date.to_local_time_string` 
+* Added bindings to `Date.to_locale_time_string_with_options` 
   [#4384](https://github.com/rustwasm/wasm-bindgen/pull/4384)
 
 ## 0.2.1